# refactor clinchedDBValues