… siteupdate/cplusplus/classes/ClinchedDBValues/ClinchedDBValues.cpp
… siteupdate/cplusplus/classes/ClinchedDBValues/ClinchedDBValues.h